pyproject.toml
[tool.poetry]
name = "epw2"
version = "0.1.0"
description = ""
authors = ["Vegenaise <moulet@proton.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-magic = {version = "^0.4.27", optional = true}
black = "^24.10.0"
py7zr = {version = "^0.22.0", optional = true}
requests = "^2.32.3"
tqdm = "^4.67.1"
colorama = "^0.4.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
